--- concert_compainion_flutter/lib/Home.dart
@@ -1,27 +1,0 @@
-import 'package:flutter/material.dart';
-
-class HomeScreen extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text("Concert Companion"),
-        actions: <Widget>[
-          IconButton(
-            icon: Icon(Icons.people),
-            onPressed: () {
-              Navigator.pushNamed(context, '/swipe');
-            },
-          ),
-          // action button
-          IconButton(
-            icon: Icon(Icons.message),
-            onPressed: () {
-              Navigator.pushNamed(context, '/message');
-            },
-          ),
-        ],
-      ),
-    );
-  }
-}
