+++ concert_compainion_flutter/pubspec.yaml
@@ -19,11 +19,12 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  permission_handler: '^4.2.0+hotfix.3'
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^0.1.2
+
+  permission_handler:
 
 dev_dependencies:
   flutter_test:
